LayoutTests/http/wpt/service-workers/navigation-fetch-worker-terminate.https.html

<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
</head>
<body>
<script>
var scope = "resources";
var activeWorker;

promise_test(async (test) => {
    var registration = await navigator.serviceWorker.register("navigation-fetch-worker-terminate-worker.js", { scope : scope });
    activeWorker = registration.active;
    if (activeWorker)
        return;
    activeWorker = registration.installing;
    return new Promise(resolve => {
        activeWorker.addEventListener('statechange', () => {
            if (activeWorker.state === "activated")
                resolve();
        });
    });
}, "Setup worker");

promise_test(async (test) => {
    const iframe = await with_iframe(scope);
    assert_equals(iframe.contentWindow.document.body.innerHTML, "PASS");
}, "Make sure a navigation load answered by a synthetic response succeeds even if worker terminates");
</script>
</body>
</html>
